Creaeted API for upload, parse and save curriculum (#11)
* renamed: schemas/file.py -> schemas/curriculum.py; ParsingException.py -> parsing_exception.py

* removed: modules/excel_parser/save_spreadsheet_data.py and modules/excel_parser/spreadsheet_schemas.py

* fix: fixed excel parser module

* renamed: renamed excel_parse on curriculum_parser

* fix: change abs import path on relative in curriculum_parser module

* fix: change return value in create methods

* fix: change return value from id model on model in add_one method

* fix: changed imperative mapping to declarative mapping style in many-to-many tables

* feat: added education_components_study_groups model to UOW

* fix: changed in to_read_model method type for semester_id on str

* fix: deleted many-to-many relationship and changed education_degree enum

* fix: changed reporting_type enum

* feat: created service and schema for EducationComponentsStudyGroups model

* fix: added optional for reporting type

* fix: changed enums value

* fix: changed credits type from int to float

* removed: remove file API

* feat: created CurriculumNotFoundException

* feat: added to __init__ files EducationComponentsStudyGroups model and repo

* feat: created utils for curriculum

* feat: created schemas for curriculum

* feat: created service for curriculum

* feat: created API for curriculum

+++ src/schemas/education_component.py
@@ -7,7 +7,7 @@ class EducationComponentBase(BaseModel):
     education_component_name: str = Field(..., max_length=255)
     education_component_code: str = Field(..., max_length=30)
     education_degree: EducationDegreeEnum
-    credits: int = Field(..., gt=0, le=100)
+    credits: float = Field(..., gt=0, le=100)
     hours: int = Field(..., ge=1, le=1000)
     department_id: str
     specialization_id: str
